requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2.1
django-material-admin==1.8.6
django-paypal==2.0
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
